resolve comments regarding noise makers
