+++ setup.py
@@ -2,8 +2,8 @@ from setuptools import setup, find_packages
 
 setup(
     name="sonosco",
-    description="Framework for training automatic speech recognition systems.",
-    author="The Roboy Gang",
+    description="Framework for deep automatic speech recognition systems.",
+    author="Roboy",
     packages=["sonosco"],
     include_package_data=True,
     dependency_links=[]
