+++ sonosco/common/audio_tools.py
@@ -2,6 +2,8 @@ import subprocess
 import numpy as np
 import librosa
 
+from .noise_makers import NoiseMaker, GaussianNoiseMaker
+
 
 def get_duration(file_path):
     return float(subprocess.check_output([f'soxi -D "{file_path.strip()}"'], shell=True))
@@ -17,23 +19,6 @@ def transcode_recordings_an4(raw_path, wav_path, sample_rate):
 
 def transcode_recordings_ted3(source, destination, start_time, end_time, sample_rate):
     subprocess.call([f"sox {source}  -r {sample_rate} -b 16 -c 1 {destination} trim {start_time} ={end_time}"],shell=True)
-
-
-class NoiseMaker:
-
-    def __call__(self, audio):
-        """Adds noise to the audio signal."""
-        pass
-
-
-class GaussianNoiseMaker(NoiseMaker):
-
-    def __init__(self, std=0.002):
-        self.std = std
-
-    def __call__(self, audio):
-        noise = np.random.randn(len(audio))
-        return audio + self.std * noise
 
 
 def add_noise(audio, noise_maker: NoiseMaker = None):
